+++ test/object-store/c_api/c_api.cpp
@@ -392,7 +392,8 @@ TEST_CASE("C API (non-database)") {
             check_mode(RLM_SCHEMA_MODE_AUTOMATIC);
             check_mode(RLM_SCHEMA_MODE_IMMUTABLE);
             check_mode(RLM_SCHEMA_MODE_READ_ONLY);
-            check_mode(RLM_SCHEMA_MODE_RESET_FILE);
+            check_mode(RLM_SCHEMA_MODE_SOFT_RESET_FILE);
+            check_mode(RLM_SCHEMA_MODE_HARD_RESET_FILE);
             check_mode(RLM_SCHEMA_MODE_ADDITIVE_EXPLICIT);
             check_mode(RLM_SCHEMA_MODE_ADDITIVE_DISCOVERED);
             check_mode(RLM_SCHEMA_MODE_MANUAL);